Move everything to a new GameObject class.
No API changes at this point.
